requirements.txt
tensorflow==2.15.0
flask==3.0.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
